+++ clj/test/limabean/adapter/beanfile_test.clj
@@ -2,7 +2,8 @@
   (:require [limabean.adapter.beanfile :as sut]
             [clojure.java.io :as io]
             [clojure.test :refer [deftest is testing]]
-            [limabean.app-test :as app-test]))
+            [limabean.app-test :as app-test]
+            [matcho.core :as matcho]))
 
 (deftest beanfile-tests
   (doseq [{:keys [name beanfile golden-dir]} (app-test/get-tests)]
@@ -11,4 +12,4 @@
             expected-directives (io/file golden-dir "directives.edn")]
         (when (.exists expected-directives)
           (let [expected (sut/read-edn-string (slurp expected-directives))]
-            (is (= (:directives actual) expected))))))))
+            (matcho/assert (:directives actual) expected)))))))