+++ rust/limabean/src/bin/limabean-pod/format.rs
@@ -1,4 +1,8 @@
+use crate::{book::types::*, plugins::InternalPlugins};
 use std::fmt::{self, Display, Formatter};
+
+#[derive(Clone, Debug)]
+struct DirectiveWithPlugins<'a, 'b>(&'b Directive<'a>, &'b InternalPlugins);
 
 // this one is subtly different from the one in beancount-parser-lima
 pub(crate) fn format<C, T, F>(
@@ -70,6 +74,7 @@ where
 pub(crate) const EMPTY: &str = "";
 pub(crate) const SPACE: &str = " ";
 pub(crate) const NEWLINE: &str = "\n";
+pub(crate) const DOUBLE_NEWLINE: &str = "\n\n";
 pub(crate) const NEWLINE_INDENT: &str = "\n  ";
 
 pub(crate) const GUTTER_MINOR: &str = " ";